requirements.txt
flask
psycopg2
Flask-Migrate
Flask-SQLAlchemy
python-dotenv
